requirements.txt
django
whitenoise
Pillow
python-decouple
